Added login functionality

+++ routes/api.js
@@ -2,6 +2,7 @@ var express = require('express');
 /* GET home page. */
 const dbInit = require('./../database/init');
 const { ObjectID } = require('mongodb');
+const crypto = require('crypto');
 
 var router = express.Router();
 
@@ -26,6 +27,8 @@ router.get('/extension/:extensionId', async function (req, res, next) {
 
 router.delete('/extension/:extensionId', async function (req, res, next) {
     console.log(`DELETE /extension/${req.params.extensionId}`);
+    const user = await validateToken(req, res);
+    if (!user) return;
     const db = await dbInit.connect();
     const result = await db.collection('extLib').deleteOne({
         _id: new ObjectID(req.params.extensionId)
@@ -42,6 +45,9 @@ router.delete('/extension/:extensionId', async function (req, res, next) {
 
 router.post('/extension/:extensionId', async function (req, res, next) {
     console.log(`POST /extension/${req.params.extensionId}`);
+    const user = await validateToken(req, res);
+    if (!user) return;
+
     const extension = req.body;
 
     const db = await dbInit.connect();
@@ -64,6 +70,9 @@ router.post('/extension/:extensionId', async function (req, res, next) {
 
 router.put('/extension/:extensionId', async function (req, res, next) {
     console.log(`PUT /extension/${req.params.extensionId}`);
+    const user = await validateToken(req, res);
+    if (!user) return;
+
     const extension = req.body;
     console.log(extension);
     if (!extension.hasOwnProperty('extension') || extension.extension == '') {
@@ -125,6 +134,9 @@ router.get('/extension', async function (req, res, next) {
 })
 
 router.post('/extension', async function (req, res, next) {
+    const user = await validateToken(req, res);
+    if (!user) return;
+
     const extension = req.body;
     if (!extension.hasOwnProperty('extension') || extension.extension == '') {
         res.status(400);
@@ -153,5 +165,71 @@ router.post('/extension', async function (req, res, next) {
     }
 
 })
+
+router.post('/token', async function (req, res, next) {
+    console.log(`POST /token`);
+    const
+        userName = req.body.userName,
+        userPassword = req.body.password,
+        userPasswordMd5 = crypto.createHash('md5').update(userPassword).digest('hex');
+
+    console.log(userName, userPasswordMd5)
+
+    const db = await dbInit.connect();
+
+    try {
+
+        const user = await db.collection('users').findOne({ userName });
+        console.log(user);
+
+        if (!user) {
+            res.status(404);
+            res.send();
+            return;
+        }
+
+        if (user.password.toLowerCase() !== userPasswordMd5.toLowerCase()) {
+            res.status(403);
+            res.send();
+            return;
+        }
+
+        let token = user.token;
+
+        if (!token) {
+            token = crypto.randomBytes(64).toString('hex');
+            console.log('NEW TOKEN', token)
+            await db.collection('users').updateOne({ _id: user._id }, { $set: { token } });
+        }
+
+        res.status(200);
+        res.send({ token });
+
+    } catch (e) {
+        console.error(e);
+        res.status(500);
+        res.send();
+        return;
+    }
+
+})
+
+async function getUserForToken(token) {
+    if (!token)
+        return false;
+    const db = await dbInit.connect();
+    const user = await db.collection('users').findOne({ token });
+    return user;
+}
+
+async function validateToken(req, res) {
+    const user = await getUserForToken(req.query.token);
+    if (!user) {
+        res.status(403);
+        res.send();
+        return false;
+    }
+    return user;
+}
 
 module.exports = router;
